App runs on both iOS and android
All content is now in `app/index.js`.
The two index.<PLATFORM>.js files just imports that one file.
