+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.pooper",
+    package = "com.poopr",
 )
 
 android_resource(
     name = "res",
-    package = "com.pooper",
+    package = "com.poopr",
     res = "src/main/res",
 )
