+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.pooper"
+    package="com.poopr"
     android:versionCode="1"
     android:versionName="1.0">
 
@@ -27,6 +27,9 @@
         </intent-filter>
       </activity>
       <activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
+      <meta-data
+        android:name="com.google.android.geo.API_KEY"
+        android:value="AIzaSyA7AwyM5SgAW15n3zTieBLKiSW5uV6hNu4"/>
     </application>
 
 </manifest>
