+++ android/app/src/main/java/com/poopr/MainActivity.java
@@ -1,4 +1,4 @@
-package com.pooper;
+package com.poopr;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "pooper";
+        return "poopr";
     }
 }
